Hierarchy serialization ... Now works! I know I said in the last commit that it *should* be working and I just didn't test it, but silly old me forgot to actually write *any* of the necessary code. But it works now. Huzzah!
Additionally, Levels are no longer `.gitignore`'d, and the Material Editor now displays the relavent textures. Also, you can ascend *above* the DataModel in the hierarchy view now, and it lists everything in `s_WorldArray` that doesn't have a parent (but of course it's just a lazy `if-condition so it's probably very easy to circumvent). And, integers can now be edited with the UI again.

Properties and functions can now be gotten statically with the 2 new `s_GetProperties` and `s_GetFunctions`. The reason the object-specific ones exist is because they can be called from `GameObject*` pointers as they are `virtual`. Also means temporary objects (that used to leak) are no longer needed for the express purpose of fetching their API for `REFLECTION_INHERITAPI`.

In case you manage to delete the DataModel with the Editor, it now shuts down gracefully instead of crashing.

Minor code refactoring. eplaced explicit `Member = DefaultValue`'s for generic `Member{}`s instead. IMPORTANTLY, instances of "ImGui" have been replaced with "Dear ImGui" for clarification and :3.