Finish GLFW migration, Luau, Editor GLFW should now be working mostly
`engine.showmessagebox` has a different signature now, it takes in the button options, default button option to select, and returns the selected optoin as a number
Removed `fs.ispromptactive`. `fs.promptopen` and `fs.promptsave` are now fully-blocking, thus `.ispromptactive` is no longer needed

The Editor will ask if you want to save the open project instead of just prompting a save dialog

I didn't fully intend for it to, but the Engine now starts at the display resolution

Bumped submodules

The Tracy Profiler cannot be built on the latest version of Ubuntu at the moment, the source code needs to be patched. See here:
https://github.com/wolfpld/tracy/pull/1142#issuecomment-3220915937

Unfortunately, for 77ccd9cb85341b735073156b24fc2a706e484202 `git diff` only showed the `miniaudio` submodule being changed, as that was the staged change, even though I use `git commit -a` to actually perform the commit, which also commits *unstaged* changes. Thus, the commit message didn't actually say that "Hey, this is the commit that removes SDL and replaced it with GLFW and tinyfiledialogs~" Oh well.